// shared/schemas.d.ts
export declare const copyRequestSchema: z.ZodObject<{
    id: z.ZodOptional<z.ZodString>;
    record_type: z.ZodString;
    reference_ids: z.ZodNullable<z.ZodOptional<z.ZodArray<z.ZodString, "many">>>;
    requester_name: z.ZodString;
    requester_cin: z.ZodString;
    request_date: z.ZodString;
    status: z.ZodDefault<z.ZodEnum<["pending", "processed"]>>;
    notes: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    primary_court: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    record_year: z.ZodNullable<z.ZodOptional<z.ZodNumber>>;
    record_details: z.ZodNullable<z.ZodOptional<z.ZodRecord<z.ZodString, z.ZodUnknown>>>;
    assigned_notary_ids: z.ZodNullable<z.ZodOptional<z.ZodArray<z.ZodString, "many">>>;
    routing_mode: z.ZodNullable<z.ZodOptional<z.ZodEnum<["direct", "historical_selection"]>>>;
}, "strip", z.ZodTypeAny, {
    id?: string;
    status?: "pending" | "processed";
    record_type?: string;
    notes?: string;
    reference_ids?: string[];
    requester_name?: string;
    requester_cin?: string;
    request_date?: string;
    primary_court?: string;
    record_year?: number;
    record_details?: Record<string, unknown>;
    assigned_notary_ids?: string[];
    routing_mode?: "direct" | "historical_selection";
}, {
    id?: string;
    status?: "pending" | "processed";
    record_type?: string;
    notes?: string;
    reference_ids?: string[];
    requester_name?: string;
    requester_cin?: string;
    request_date?: string;
    primary_court?: string;
    record_year?: number;
    record_details?: Record<string, unknown>;
    assigned_notary_ids?: string[];
    routing_mode?: "direct" | "historical_selection";
}>;
// ...
export type CopyRequest = z.infer<typeof copyRequestSchema>;